backend/requirements.txt
fastapi
uvicorn
phidata
psycopg2-binary
loguru
python-dotenv
transformers
torch
neo4j
pymilvus
pymongo
pydantic
minio
etcd3
openai
dagster
dagit
aiohttp
requests
yfinance
stable-baselines3
tensorboard
protobuf==3.20.3
cachetools==5.3.1  
yfinance==0.2.40  
scikit-learn
stable-baselines3
backtrader
python-multipart
beautifulsoup4




